refactor: version > policy 로 이름 변경
