+++ src/main/java/com/groupMeeting/core/exception/custom/PolicyException.java
@@ -4,10 +4,10 @@ import com.groupMeeting.global.enums.ExceptionReturnCode;
 import lombok.Getter;
 
 @Getter
-public class VersionException extends RuntimeException {
+public class PolicyException extends RuntimeException {
     private final ExceptionReturnCode exceptionReturnCode;
 
-    public VersionException(ExceptionReturnCode exceptionReturnCode) {
+    public PolicyException(ExceptionReturnCode exceptionReturnCode) {
         super(exceptionReturnCode.getMessage());
         this.exceptionReturnCode = exceptionReturnCode;
     }
