+++ src/main/java/com/groupMeeting/policy/repository/ApiVersionPolicyRepository.java
@@ -1,6 +1,6 @@
-package com.groupMeeting.version.repository;
+package com.groupMeeting.policy.repository;
 
-import com.groupMeeting.entity.version.ApiVersionPolicy;
+import com.groupMeeting.entity.policy.ApiVersionPolicy;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ApiVersionPolicyRepository extends JpaRepository<ApiVersionPolicy, Long> {
